Bluetooth: Mesh: Get rid of unused scene entries
Scene entry is defined and initialized by every model that has a state
stored with scene. However, when a model is extended by another model,
the Scene Store and Scene Recall behaviours are controled by the
extending model. Thus, scene_store and scene_recall will never be called
for the extended model. This change makes possible for the compiler to
get rid of unused code.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>