sysbuild: Support Bluetooth Mesh default config for netcore
There are few issues with this approach:
1. If SB_CONFIG_NETCORE_BT_MESH_DEFAULTS is enabled and user defines
   extended advertising API configuration in
   `<app>/sysbuild/hci_ipc/prj.conf`, this configuration will be
   overwritten by the overlay file. A user must either deselect this
   option or use own overlay in `<app>/sysbuild/hci_ipc.conf`.
2. The mesh overlay file must be injected first, otherwise it will be
   injceted after user's overlay file and will overwrite user's
   configuration.
3. For samples, SB_CONFIG_NETCORE_BT_MESH_DEFAULTS must be enabled in
   `<app>/Kconfig.sysbuild`. If it is enabled in `<app>/sysbuild.conf`,
   it will generate warning on nRF52 because it depends on netcore
   Kconfig, which is not present on nRF52. There will be no
   `<app>/sysbuild_<board>.conf` support in sysbuild.
4. We need to mention somewhere in docs this new option and how to use
   it (see previous points). We also become dependent on it if we find
   better way of handling netcore configuration for mesh (NCSDK-18316).
5. Most likely customers will clone any of our samples and will meet the
   issues described above until they figure out how to work with this
   option.

I feel we need to add a way to enable mesh defaults on netcore through
subsys Kconfig files instead of using this approach. See NCSDK-18316.

------------------------------------------------------------------------

The Bluetooth Mesh default configuration in nRF Connect SDK requires the
extended advertising API and specific number of advertising sets enabled
on the Bluetooth Controller. When the host and controller are split
between cores, this configuration is not passed to the network core.
Due to misconfiguration, the mesh app can't start.

With the child image, we used an overlay file to pass the default
configuration to the network core. This commit allows to enable this
configuration on the network core with sysbuild.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>