Managed to set variable name through dts file
Now I can provide variable name of a model in dts. This allows to move
this macrobiotic into a header file.

`cmodel` is used to convert the value into a model macro <model>_MACRO,
which takes `vname` as an argument. If model requires a model context
(e.g. `bt_mesh_onoff_cli` struct), the name is used to pass it to model
macro. And the context must be initialized by a user somewhere in the
code next to element initialization.

Created bt_mesh_element.yaml and bt_mesh_model.yaml to specify required
properties, thought it is possible to embedd bt_mesh_model.yaml into
bt_mesh_element.yaml (see commented code).

It doesn't seem to be possible to require models lists node in dts...